+++ CHANGELOG.md
@@ -7,10 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [v1.2.1]
+
 ### Add
 - Print special abilities option in config
 - Colors to cards
 - 'All' button for year selector
+- Save list and load options
 
 ### Edit
 - Move admin menu to footer
@@ -40,7 +43,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Version 1.0.0 published!
 
-[Unreleased]: https://github.com/juananmuxed/air-raid-builder/compare/v1.2.0...HEAD
+[Unreleased]: https://github.com/juananmuxed/air-raid-builder/compare/v1.2.1...HEAD
+[v1.2.1]: https://github.com/juananmuxed/air-raid-builder/releases/tag/v1.2.1
 [v1.2.0]: https://github.com/juananmuxed/air-raid-builder/releases/tag/v1.2.0
 [v1.0.1]: https://github.com/juananmuxed/air-raid-builder/releases/tag/v1.0.1
 [v1.0.0]: https://github.com/juananmuxed/air-raid-builder/releases/tag/v1.0.0